run several population simulations

run several population simulations and visualize the result, refer to vignette('pgmcs') for examples

demo_population_simulation(nrun, ngeneration, population, fitness, title)

Arguments

nrun

integer scalar. Number of simulation runs.

ngeneration

integer scalar. Number of generations

population

integer vector. How many dominant homozygotes, heterozygotes, and recessive homozygotes do the initial population contain. For example, c(2, 5, 3) means 2 AA, 5 Aa and 3 aa.

fitness

numeric vector. The fitness value of AA, Aa and aa. c(1, 1, 1) means no natural selection.

title

string. Plot title.

Value

an ggplot object

Details

the underlying workhorse is population_simulation()

Shiny app

shiny::runApp(system.file('shiny', package = 'pgs'))